Move merkle/common to NodeID
