+++ merkle/common.go
@@ -24,27 +24,3 @@ func parent(leafIndex int64) int64 {
 func isRightChild(leafIndex int64) bool {
 	return leafIndex&1 == 1
 }
-
-// bit returns the i'th bit of index from the right.
-// eg. bit(0x80000000, 31) -> 1
-func bit(index []byte, i int) uint {
-	IndexBits := len(index) * 8
-	bIndex := (IndexBits - i - 1) / 8
-	return uint((index[bIndex] >> uint(i%8)) & 0x01)
-}
-
-// flipBit returns index with the i'th bit from the right flipped.
-func flipBit(index []byte, i int) []byte {
-	r := make([]byte, len(index))
-	copy(r, index)
-	IndexBits := len(index) * 8
-	bIndex := (IndexBits - i - 1) / 8
-	r[bIndex] ^= 1 << uint(i%8)
-	return r
-}
-
-// Neighbor returns index with only the left i bits set and the i'th bit flipped.
-func Neighbor(index []byte, i int) []byte {
-	r := flipBit(index, i)
-	return MaskIndex(r, len(index)*8-i)
-}
