+++ storage/types_test.go
@@ -303,6 +303,31 @@ func TestSetBit(t *testing.T) {
 	}
 }
 
+func TestFlipBit(t *testing.T) {
+	for _, tc := range []struct {
+		index []byte
+		i     int
+		want  []byte
+	}{
+		{index: h2b("00"), i: 0, want: h2b("01")},
+		{index: h2b("00"), i: 7, want: h2b("80")},
+		{index: h2b("000b"), i: 0, want: h2b("000a")},
+		{index: h2b("000b"), i: 1, want: h2b("0009")},
+		{index: h2b("000b"), i: 2, want: h2b("000f")},
+		{index: h2b("000b"), i: 3, want: h2b("0003")},
+		{index: h2b("0001"), i: 0, want: h2b("0000")},
+		{index: h2b("8000"), i: 15, want: h2b("0000")},
+		{index: h2b("0000000000000001"), i: 0, want: h2b("0000000000000000")},
+		{index: h2b("0000000000010000"), i: 16, want: h2b("0000000000000000")},
+		{index: h2b("8000000000000000"), i: 63, want: h2b("0000000000000000")},
+	} {
+		nID := NewNodeIDFromHash(tc.index)
+		if got, want := nID.FlipRightBit(tc.i).Path, tc.want; !bytes.Equal(got, want) {
+			t.Errorf("flipBit(%x, %d): %x, want %x", tc.index, tc.i, got, want)
+		}
+	}
+}
+
 func TestBit(t *testing.T) {
 	for _, tc := range []struct {
 		n    NodeID
